refactor: create utils/constants and update import
