+++ apps/mobile/src/screens/Dashboard/History/index.tsx
@@ -1,12 +1,10 @@
 import type { FC } from 'react';
 import type { ViewStyle } from 'react-native';
 import type { StackScreenProps } from '@react-navigation/stack';
-import {
-	FullHistoryFeature,
-	tabBarHeight,
-	useSafeAreaInsets,
-} from '@walless/app';
+import { FullHistoryFeature } from '@walless/app';
 import { View } from '@walless/gui';
+import { tabBarHeight } from 'utils/constants';
+import { useSafeAreaInsets } from 'utils/hooks';
 import { type HomeParamList } from 'utils/navigation';
 
 type Props = StackScreenProps<HomeParamList, 'History'>;
